requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
psycopg2-binary
pydantic
pydantic-settings
requests
typer
pytest
httpx
rich
python-jose[cryptography]
passlib[bcrypt]
bcrypt==3.2.0
python-multipart
python-dotenv
